backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
stream-chat==4.20.0
openai==1.82.0
ultralytics==8.3.145
deep-sort-realtime==1.3.2
numpy==2.2.6
opencv-python-headless==4.11.0.86
Pillow==11.2.1
pydantic==2.11.4
pytest==8.3.5
mediapipe==0.10.14
scikit-learn==1.5.2
